requirements.txt
streamlit
networkx
google-generativeai
streamlit-agraph
pandas
plotly
